refactor: remove GenaiLangChainProcessor import from contrib __init__.py for modularity
