+++ genai_processors/contrib/__init__.py
@@ -13,4 +13,3 @@
 # limitations under the License.
 # ==============================================================================
 """Community contributed processors."""
-from .genai_langchain_streaming import GenAILangChainProcessor
